# Conflicts:
#	CHANGELOG.md

+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/util/ManualFileDescriptor.kt
@@ -5,7 +5,7 @@ import net.folivo.trixnity.utils.ByteArrayFlow
 
 class ManualFileDescriptor(
     override val fileName: String,
-    override val fileSize: Int?,
+    override val fileSize: Long?,
     override val mimeType: ContentType?,
     override val content: ByteArrayFlow
 ) : FileDescriptor {}
